MNT: split widgets/core.py into individual modules
