+++ docs/source/widgets/index.rst
@@ -1,9 +1,6 @@
 =======
 Widgets
 =======
-.. automodule:: squirrel.widgets.core
-   :members:
-
 .. automodule:: squirrel.widgets.flow_layout
    :members:
 
@@ -14,4 +11,19 @@ Widgets
    :members:
 
 .. automodule:: squirrel.widgets.window
+   :members:
+
+.. automodule:: squirrel.widgets.namedesctags
+   :members:
+
+.. automodule:: squirrel.widgets.qsingleton
+   :members:
+
+.. automodule:: squirrel.widgets.data_widget
+   :members:
+
+.. automodule:: squirrel.widgets.display
+   :members:
+
+.. automodule:: squirrel.widgets.window_linker
    :members:
