+++ squirrel/tests/test_widgets.py
@@ -14,8 +14,7 @@ from squirrel.control_layer import EpicsData
 from squirrel.model import Collection, Snapshot
 from squirrel.tables import PVTableModel
 from squirrel.tests.conftest import setup_test_stack
-from squirrel.widgets import TagsWidget
-from squirrel.widgets.core import DataWidget
+from squirrel.widgets import DataWidget, TagsWidget
 
 
 @pytest.fixture(scope='function')
