--- squirrel/widgets/core.py
@@ -1,250 +1,0 @@
-"""
-Core widget classes for qt-based GUIs.
-"""
-from __future__ import annotations
-
-from pathlib import Path
-from typing import ClassVar, Optional
-from weakref import WeakValueDictionary
-
-from pcdsutils.qt.designer_display import DesignerDisplay
-from qtpy import QtCore, QtWidgets
-
-from squirrel.client import Client
-from squirrel.qt_helpers import QDataclassBridge
-from squirrel.type_hints import AnyDataclass, OpenPageSlot, TagDef
-from squirrel.utils import SQUIRREL_SOURCE_PATH
-from squirrel.widgets import TagsWidget, get_window
-
-
-class Display(DesignerDisplay):
-    """Helper class for loading designer .ui files and adding logic"""
-
-    ui_dir: Path = SQUIRREL_SOURCE_PATH / 'ui'
-
-
-class DataWidget(QtWidgets.QWidget):
-    """
-    Base class for widgets that manipulate dataclasses.
-
-    Defines the init args for all data widgets and handles synchronization
-    of the ``QDataclassBridge`` instances. This is done so that only data
-    widgets need to consider how to handle bridges and the page classes
-    simply need to pass in data structures, rather than needing to keep track
-    of how two widgets editing the same data structure must share the same
-    bridge object.
-
-    Parameters
-    ----------
-    data : any dataclass
-        The dataclass that the widget needs to manipulate. Most widgets are
-        expecting either specific dataclasses or dataclasses that have
-        specific matching fields.
-    kwargs : QWidget kwargs
-        Passed directly to QWidget's __init__. Likely unused in most cases.
-        Even parent is unlikely to see use because parent is set automatically
-        when a widget is inserted into a layout.
-    """
-    # QDataclassBridge for this widget, other bridges may live in EntryItem
-    _bridge_cache: ClassVar[
-        WeakValueDictionary[int, QDataclassBridge]
-    ] = WeakValueDictionary()
-    bridge: QDataclassBridge
-    data: AnyDataclass
-
-    def __init__(self, data: AnyDataclass, **kwargs):
-        super().__init__(**kwargs)
-        self.data = data
-        try:
-            # TODO figure out better way to cache these
-            # TODO worried about strange deallocation timing race conditions
-            self.bridge = self._bridge_cache[id(data)]
-        except KeyError:
-            bridge = QDataclassBridge(data)
-            self._bridge_cache[id(data)] = bridge
-            self.bridge = bridge
-
-
-class NameMixin:
-    """
-    Mixin class for distributing init_name
-    """
-    def init_name(self) -> None:
-        """
-        Set up the name_edit widget appropriately.
-        """
-        # Load starting text
-        load_name = self.bridge.title.get() or ''
-        self.last_name = load_name
-        self.name_edit.setText(load_name)
-        # Set up the saving/loading
-        self.name_edit.textEdited.connect(self.update_saved_name)
-        self.bridge.title.changed_value.connect(self.apply_new_name)
-
-    def update_saved_name(self, name: str) -> None:
-        """
-        When the user edits the name, write to the config.
-        """
-        self.last_name = self.name_edit.text()
-        self.bridge.title.put(name)
-
-    def apply_new_name(self, text: str) -> None:
-        """
-        If the text changed in the data, update the widget.
-
-        Only run if needed to avoid annoyance with cursor repositioning.
-        """
-        if text != self.last_name:
-            self.name_edit.setText(text)
-
-
-class NameDescTagsWidget(Display, NameMixin, DataWidget):
-    """
-    Widget for displaying and editing the name, description, and tags fields.
-
-    Any of these will be automatically disabled if the data source is missing
-    the corresponding field.
-    """
-    filename = 'name_desc_tags_widget.ui'
-
-    name_edit: QtWidgets.QLineEdit
-    name_frame: QtWidgets.QFrame
-    desc_edit: QtWidgets.QPlainTextEdit
-    desc_frame: QtWidgets.QFrame
-    tags_widget: TagsWidget
-
-    def __init__(self, data: AnyDataclass, **kwargs):
-
-        tag_groups = kwargs.pop('tag_options', dict())
-
-        super().__init__(data=data, **kwargs)
-        try:
-            self.bridge.title
-        except AttributeError:
-            self.name_frame.hide()
-        else:
-            self.init_name()
-        try:
-            self.bridge.description
-        except AttributeError:
-            self.desc_frame.hide()
-        else:
-            self.init_desc()
-        try:
-            self.bridge.tags
-        except AttributeError:
-            self.tags_widget.hide()
-        else:
-            self.init_tags(tag_groups)
-
-    def init_desc(self) -> None:
-        """
-        Set up the desc_edit widget appropriately.
-        """
-        # Load starting text
-        load_desc = self.bridge.description.get() or ''
-        self.last_desc = load_desc
-        self.desc_edit.setPlainText(load_desc)
-        # Setup the saving/loading
-        self.desc_edit.textChanged.connect(self.update_saved_desc)
-        self.bridge.description.changed_value.connect(self.apply_new_desc)
-
-    def update_saved_desc(self) -> None:
-        """
-        When the user edits the desc, write to the config.
-        """
-        self.last_desc = self.desc_edit.toPlainText()
-        self.bridge.description.put(self.last_desc)
-
-    def apply_new_desc(self, desc: str) -> None:
-        """
-        When some other widget updates the description, update it here.
-        """
-        if desc != self.last_desc:
-            self.desc_edit.setPlainText(desc)
-
-    def init_tags(self, tag_groups: TagDef) -> None:
-        """
-        Set up the tags widgets appropriately.
-        """
-        self.tags_widget.setObjectName("TagsWidget")
-        self.tags_widget.setEnabled(True)
-        self.tags_widget.set_tag_groups(tag_groups)
-
-
-class QtSingleton(type(QtCore.QObject), type):
-    """
-    Qt specific singleton implementation, needed to ensure signals are shared
-    between instances.  Adapted from
-    https://stackoverflow.com/questions/59459770/receiving-pyqtsignal-from-singleton
-
-    The more common __new__ - based singleton pattern does result in the QObject
-    being a singleton, but the bound signals lose their connections whenever the
-    instance is re-acquired.  I do not understand but this works
-
-    To use this, specify `QtSingleton` as a metaclass:
-
-    .. code-block:: python
-
-        class SingletonClass(QtCore.QObject, metaclass=QtSingleton):
-            shared_signal: ClassVar[QtCore.Signal] = QtCore.Signal()
-
-    """
-    def __init__(cls, name, bases, dict):
-        super().__init__(name, bases, dict)
-        cls._instance = None
-
-    def __call__(cls, *args, **kwargs):
-        if cls._instance is None:
-            cls._instance = super().__call__(*args, **kwargs)
-        return cls._instance
-
-
-class WindowLinker:
-    """
-    Mixin class that provides access methods for resources held by the main Window.
-    These include:
-    - client: first attempts to grab the client set at init, if none exists use the Window's client
-    - open_page_slot: grabs the slot from the Window
-    """
-
-    def __init__(self, *args, client: Optional[Client] = None, **kwargs) -> None:
-        self._client = client
-        super().__init__(*args, **kwargs)
-
-    @property
-    def client(self) -> Optional[Client]:
-        # Return the provided client if it exists, grab the Window's otherwise
-        if self._client is not None:
-            return self._client
-        else:
-            window = get_window()
-            if window is not None:
-                return window.client
-
-    @client.setter
-    def client(self, client: Client):
-        if not isinstance(client, Client):
-            raise TypeError(f"Cannot set a {type(client)} as a client")
-
-        if client is self._client:
-            return
-
-        self._client = client
-
-    @property
-    def open_page_slot(self) -> Optional[OpenPageSlot]:
-        window = get_window()
-        if window is not None:
-            return window.open_page
-
-    def get_window(self):
-        """Return the singleton Window instance"""
-        return get_window()
-
-    def refresh_window(self):
-        """Refresh window ui elements"""
-        # tree view
-        window = get_window()
-        window.tree_view.set_data(self.client.backend.root)
-        window.tree_view.model().refresh_tree()
